Add x86 vendor conversions
